requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
dnspython==2.3.0
pkg-resources==0.0.0
pymongo==4.7.3
python-decouple==3.8
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.7.1
